pyproject.toml
[build-system]
build-backend = "hatchling.build"

requires = [ "hatchling" ]

[project]
name = "fastapi-ronin"
version = "0.3.1"
description = "FastAPI Ronin is a library for building FastAPI applications with a Django REST Framework-inspired architecture."
readme = "README.md"
keywords = [
    "django",
    "django-rest-framework",
    "fastapi",
    "pagination",
    "tortoise",
    "viewsets",
]
license = { text = "MIT" }
maintainers = [ { name = "bubaley", email = "bubaley.fu@gmail.com" } ]
authors = [ { name = "bubaley", email = "bubaley.fu@gmail.com" } ]
requires-python = ">=3.12"
classifiers = [
    "Development Status :: 4 - Beta",
    "Framework :: FastAPI",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Internet :: WWW/HTTP",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Typing :: Typed",
]

dependencies = [
    "fastapi>=0.115.13",
    "tortoise-orm>=0.25.1",
]

urls.Documentation = "https://github.com/bubaley/fastapi-ronin#readme"
urls.Homepage = "https://github.com/bubaley/fastapi-ronin"
urls.Repository = "https://github.com/bubaley/fastapi-ronin.git"

[dependency-groups]
dev = [
    "aerich[toml]>=0.9.1",
    "coverage>=7.9.2",
    "pre-commit>=4.2.0",
    "pydantic-settings>=2.10.1",
    "ruff>=0.12.2",
    "uvicorn[standard]>=0.35.0",
]
docs = [
    "mkdocs>=1.5.0",
    "mkdocs-material>=9.0.0",
    "mkdocstrings[python]>=0.20.0",
    "pymdown-extensions>=10.0.0",
]

[tool.ruff]
target-version = "py313"
line-length = 120
indent-width = 4
fix = true
format.quote-style = "single"
format.line-ending = "auto"
format.skip-magic-trailing-comma = false
lint.extend-select = [ "I" ]
lint.extend-per-file-ignores."core/settings/*.py" = [ "F403", "F405" ]
lint.dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.pyright]
typeCheckingMode = "basic"

[tool.bandit]
skips = [ "B101" ]
exclude_dirs = [ "tests", "*/tests/*", "*/migrations/*", "scripts/test-docs.py" ]

[tool.aerich]
tortoise_orm = "app.core.database.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."
